refactor: move ToggleButton to parent components directory
