+++ src/refactoring/pages/AdminPage.tsx
@@ -1,7 +1,7 @@
 import { Coupon, Product } from "../../types.ts";
 import { ProductEditAccordion } from "../components/admin/ProductEditAccordion.tsx";
 import { ProductAddForm } from "../components/admin/ProductAddForm.tsx";
-import { ToggleButton } from "../components/ui/ToggleButton.tsx";
+import { ToggleButton } from "../components/ToggleButton.tsx";
 import { CouponAddForm } from "../components/admin/CouponAddForm.tsx";
 import { ProductManageContainer } from "../components/admin/ProductManageContainer.tsx";
 import { CouponManageContainer } from "../components/admin/CouponManageContainer.tsx";
